requirements.txt
Flask
PyPDF2
pytesseract
Pillow
pdf2image